Added/Updated: faq.css, general.css and index.html
Added: faq.css to contain styles for the faq container, its children, as well as their
active states, media queries, and etc.
Updated general.css so that the body has the scrollbar turned on by default for consistency
and a lot of minor changes.
Updated index.html and removed the img tags for the summary tags and instead utilized CSS pseudo elements.